[AO] new feature: add custom headers field to the rest query interface

+++ app/uk/gov/hmrc/agentsexternalstubsfrontend/controllers/RestQueryController.scala
@@ -23,7 +23,7 @@ import com.google.inject.Provider
 import javax.inject.{Inject, Singleton}
 import play.api.Configuration
 import play.api.data.Form
-import play.api.data.Forms.{mapping, optional, text}
+import play.api.data.Forms.{mapping, optional, text, seq, tuple}
 import play.api.data.validation.{Constraint, Invalid, Valid}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.libs.json._
@@ -40,6 +40,8 @@ import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
+import scala.util.Try
+import java.net.URL
 
 @Singleton
 class RestQueryController @Inject()(
@@ -65,7 +67,7 @@ class RestQueryController @Inject()(
             case None =>
               Future.successful(
                 Ok(restQueryView(
-                  RestQueryForm.fill(RestQuery("GET", "https://", None)),
+                  RestQueryForm.fill(RestQuery("GET", "https://", None, None)),
                   routes.RestQueryController.runQuery(),
                   routes.UserController.showUserPage(),
                   routes.RestQueryController.showRestQueryPage(None),
@@ -134,17 +136,28 @@ class RestQueryController @Inject()(
 
   private def runQuery(
     query: RestQuery)(implicit ec: ExecutionContext, request: Request[AnyContent]): Future[WSResponse] = {
-    val wsRequest = wsClient
-      .url(query.url)
-      .withHeaders(
+    val headers = Map(
         HeaderNames.AUTHORIZATION -> request.session.get(SessionKeys.authToken).get,
-        "X-Session-ID"            -> request.session.get(SessionKeys.sessionId).get)
-    query.method match {
-      case "GET"    => wsRequest.get()
-      case "POST"   => wsRequest.post(query.payload.getOrElse(JsNull))
-      case "PUT"    => wsRequest.put(query.payload.getOrElse(JsNull))
-      case "DELETE" => wsRequest.delete()
-      case _        => Future.failed(new Exception(s"Method ${query.method} is not supported, try GET, POST, PUT or DELETE"))
+        "X-Session-ID"            -> request.session.get(SessionKeys.sessionId).get) ++ 
+        query.headers.map(RestQueryController.parseHeaders).getOrElse(Map.empty)  
+    if (Try(new URL(query.url)).isFailure) 
+      Future.failed(new Exception(s"Invalid URL ${query.url}"))
+    else {
+      Try{
+          val wsRequest = wsClient
+          .url(query.url)
+          .withHttpHeaders(headers.toSeq:_*)
+        query.method match {
+          case "GET"    => wsRequest.get()
+          case "POST"   => wsRequest.post(query.payload.getOrElse(JsNull))
+          case "PUT"    => wsRequest.put(query.payload.getOrElse(JsNull))
+          case "DELETE" => wsRequest.delete()
+          case _        => Future.failed(new Exception(s"Method ${query.method} is not supported, try GET, POST, PUT or DELETE"))
+        }
+      }.fold(
+        e => Future.failed(new Exception(s"Error executing the request: ${e.getMessage}")),
+        identity
+      )
     }
   }
 
@@ -152,20 +165,22 @@ class RestQueryController @Inject()(
 
 object RestQueryController {
 
-  case class RestQuery(method: String, url: String, payload: Option[JsValue]) {
+  case class RestQuery(method: String, url: String, payload: Option[JsValue], headers: Option[String]) {
     def encode: String =
       new String(
         Base64.getUrlEncoder.encode(Json.toJson(this).toString().getBytes(StandardCharsets.UTF_8)),
         StandardCharsets.UTF_8)
 
-    def toCurlCommand(implicit request: Request[AnyContent]): String =
-      s"""curl -v -X $method -H "Authorization: ${request.session
-        .get(SessionKeys.authToken)
-        .get}" -H "X-Session-ID: ${request.session
-        .get(SessionKeys.sessionId)
-        .get}" ${if (method == "POST" || method == "PUT") """-H "Content-Type: application/json"""" else ""} ${payload
+    def toCurlCommand(implicit request: Request[AnyContent]): String = {
+      val h = Map(
+        "Authorization"->request.session.get(SessionKeys.authToken).get,
+        "X-Session-ID" -> request.session.get(SessionKeys.sessionId).get
+      ) ++ headers.map(RestQueryController.parseHeaders).getOrElse(Map.empty)
+
+      s"""curl -v -X $method ${h.map{case (key,value) => s"""-H "$key: $value""""}.mkString(" ")} ${if (method == "POST" || method == "PUT") """-H "Content-Type: application/json"""" else ""} ${payload
         .map(p => s"""--data '$p'""")
         .getOrElse("")} $url"""
+    }
   }
 
   object RestQuery {
@@ -187,10 +202,35 @@ object RestQueryController {
   val RestQueryForm: Form[RestQuery] = Form[RestQuery](
     mapping(
       "method" -> text,
-      "url"    -> text,
+      "url"    -> text.verifying("error.url", u => Try(new URL(u)).isSuccess),
       "payload" -> optional(text)
         .verifying(validJson)
-        .transform[Option[JsValue]](_.map(Json.parse), _.map(Json.prettyPrint))
+        .transform[Option[JsValue]](_.map(Json.parse), _.map(Json.prettyPrint)),
+      "headers" -> optional(text).verifying("error.headers", _.forall{ h => parseHeaders(h); true})
     )(RestQuery.apply)(RestQuery.unapply))
+
+  def prettyPrintHeaders(headers: Seq[(String,String)]): Option[String] = 
+    if (headers.isEmpty) None 
+    else 
+      Some( headers.map { 
+        case (key, value) => s"$key: $value"
+      }.mkString("\n"))
+
+  def parseHeaders(headers: String): Map[String,String] = 
+    if (headers.isEmpty) Map.empty 
+    else 
+      headers
+        .lines
+        .map { line => 
+          val key = line.takeWhile(_ != ':')
+          if(key.nonEmpty) {
+            val value = line.drop(key.length + 1)
+            if(value.nonEmpty) {
+              Some((key.trim(), value.trim()))
+            } else None
+          } else None
+        }
+        .collect {case Some(x) => x}
+        .toMap
 
 }
